feat: update project to fit more in line with MVVM and allow for potential expansion
Closes #1
