+++ Base/ViewModelBase.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CrystalBlue.Base
+{
+    /// <summary>
+    /// Base classes for ViewModels implementing <see cref=INotifyPropertyChanged"/>
+    /// </summary>
+    public class ViewModelBase : INotifyPropertyChanged
+    {
+        /// <summary>
+        /// Property Changed Event Handler
+        /// </summary>
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        /// <summary>
+        /// Handles the raising of property changed to update the view
+        /// </summary>
+        /// <param name="name"></param>
+        public void OnPropertyChanged( [CallerMemberName] string name = null )
+        {
+            PropertyChanged?.Invoke( this, new PropertyChangedEventArgs( name ) );
+        }
+    }
+}
